requirements.txt
appdirs==1.4.4
astroid==2.4.2
attrs==19.3.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
bleach==3.1.5
boto==2.49.0
cachelib==0.1.1
certifi==2020.4.5.2
chardet==3.0.4
charset-normalizer==2.0.2
click==7.1.2
colorama==0.4.3
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
entrypoints==0.3
filelock==3.0.12
Flask==2.0.1
Flask-Session==0.3.2
Flask-Sessions==0.1.5
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.9
ipython==7.16.1
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==2.0.1
jedi==0.17.2
Jinja2==3.0.1
jsonschema==3.2.0
jupyter-core==4.6.3
lazy-object-proxy==1.4.3
MarkupSafe==2.0.1
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
packaging==20.4
pandocfilters==1.4.2
parso==0.7.1
pickleshare==0.7.5
platformdirs==2.0.2
prompt-toolkit==3.0.5
psycopg2==2.9.1
Pygments==2.6.1
pylint==2.5.3
pyparsing==2.4.7
pyrsistent==0.16.0
python-dotenv==0.15.0
requests==2.26.0
six==1.15.0
spotipy==2.13.0
testpath==0.4.4
toml==0.10.1
traitlets==4.3.3
urllib3==1.26.6
virtualenv==20.6.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
wrapt==1.12.1
WTForms==2.3.1
